♻️ Refactor and cleanup namespaces.
